Add daemon test
Add a test of the way that several zuul processes start using
the daemon and pidlockfile modules.

Ideally we would refactor this to use the actual code without
duplication, but recent releases of daemon are causing real problems
so suffer some duplication so that we can at least better evaluate
changes.

Change-Id: Ibaa0f481a52b6177dfffe1a58ebff342873df51b