Cloner: handle missing ZUUL_BRANCH/_REF
Periodic jobs currently trigger without a ZUUL_BRANCH or ZUUL_REF
variable set.  (We should change this, but that's a more complex
change).

So that the cloner can be used in this situation, set those to
the empty string if their value is None (ie, unset).

Periodoc jobs can then invoke the cloner as normal, with branch
override values set, and those will be used as the selected
branch.  This is how OpenStack's periodic branch jobs currently
work.

In the future, if Zuul itself performs branch selection on jobs,
this will provide a graceful migration path by allowing the
ZUUL_BRANCH values to act as a fallback and then the overrides
can be removed.

Add a test to validate this usage.

Change-Id: Ie680b9c11cebec7410ef7b0edc9d1017ae8b847e