Add aditional CRD tests
Test that when two changes (in different branches) share a change-id
that Zuul depends on both.

Test that when two change ids are listed in a commit, zuul depends
on both.

Test that reconfiguration with changes in the check queue still
produces the same non-live + live change pair in a shared queue.

Also update the change history assertion in some tests to verify
the 'project1-merge' job by name which is less racy than checking
the last job (which may vary).

Change-Id: Ib2b8202f24578b10aa706befcca40b85459e1b97