Fix checking all builds are waiting in tests
The check for whether all outstanding builds are waiting (which
means the system is stable and may be examined) did not account
for builds which may have just started running but had not reported
their start back to Zuul.  This means that the test system could
determine the system had settled while a job start packet was
in-flight.  This has been seen to cause the test_abandoned_gate
test to fail due to the inconsistent state.

Since this race can occur in reality, also update the launcher
itself to more gracefully handle this condition.

Change-Id: I734094514db294564f6526e42b801a7e1d22a021