Status: Expose $.zuul object in zuul_start to allow extension
When embedding this into a branded dashboard with other pages
it is preferable not to add custom hooks into these files but
instead call zuul_start and be able to attach additional event
handlers to zuul.app.on() or zuul.jq.on().

Used on https://integration.wikimedia.org/zuul/ and would likely
be needed on http://status.openstack.org/zuul/ in the future
as well.

Change-Id: I77d935d67051d50d38db5a002f4bf766c4371731