Update to gear 0.5.7
There is no API or behavioral change required, but gear 0.5.7 has
significant performance improvements, enough so that it is required
for OpenStack's installation, so bump the requirement here.

Gear 0.5.6 has a regression that causes Zuul tests (but not actual
operation) to fail.

Also includes an unrelated test race fix: the order in which jobs
completed in the test_crd_check_git_depends test was random.  Switch
to releasing them in sequence for easier comparison with the build
history at the end.

Change-Id: I281356a620c017a546ec04a0a767cf34c1fc62bf