Allow all zuul args to be set in zuul-cloner
Previously zuul-cloner required that no zuul-args be set. The intent was
that all args be set or none be set and to disallow any partial setting
of the zuul args. Add the condition to allow all zuul args to be set by
allowing the length of the missing zuul args list to be 0.

Change-Id: I944f12aa1148bce7d3998b400e8d6e86fd0164d5