Don't send reports without any job
Some CIs like to narrow their scope to a certain set of files.
For that, they specify file mask on per-job basis. So there appear
annoying comments with only "Build succeeded".
(an example complaint:
http://lists.openstack.org/pipermail/openstack-dev/2015-June/065367.html)

Moreover, most of CIs which don't bother filtering, make lots of
comments to doc/unittest changes, which is also wrong.

By this change we drop such reports, leaving a record at debug level.
We also don't want to report on failures in such configuration,
as it's impossible to recheck them.

Change-Id: Ifa43b6a2cd07d81b26dd2fe48e780c5ac94dbbb9