Set build descriptions less often
When busy, Zuul spends a lot of time keeping these up to date,
however, most users of the feature only look at this data after
builds are completed.  Therefore, only set build descriptions
for a build set when that build set is completed (when an item
is reported or reset).

Change-Id: I7c255937f08aef28bdad72e8f523ca34f2c6262c