Set build state after timing
There is a race where the result of a build may be recorded before
its end time.  As soon as a build's result is recorded, the main
thread may start using that information.  In unit tests, it has
been observed to issue a report for a change between setting the
result and setting the end time, thereby reporting a change with
no timing information.  Ensure that ancillary data about a build
are recorded before the result is to avoid this condition.

Change-Id: Idb902168bb7b89cda43c90ef9ddc9504b6bad9cc