Don't use outdated repo object in reset
When resetting a repository, the merger performs an update of the
origin remote to ensure that it has the latest HEAD.  This was done
after the reset method instantiated its own repo object.  It is
possible that the repository may have changed sufficiently via
the update method (which uses its own repo object) that the original
repo object may not function correctly.

Change-Id: I78354999be7a2ad2b113c9637b436734b6be4eb9