Specify gearman server listen address with separate option
Recently the ability for the gearman server to listen on a specific
address was added, but it simply re-used the address already in the
config file.  That might be fine for some cases, but it's strange
for the case where you want it listening on all addresses but
connecting to localhost.

Add a new listen_address for the gearman_server config section
so that the listen and connect addresses can be supplied separately.

Change-Id: Id23adb90145984e74ea4372ef4672fc06c7afa90