Fix IndexError: No item found with id u'origin/HEAD in tests
The test "test_live_cenfogirutaion_merge_conflict" introduced a
race by adding changes to the "upstream" repository while Zuul
was processing.  This could mean that the merger's initial clone
could happen while the upstream repo was in an unusual state.
Correct this by creating all of the changes and patchsets while
Zuul is idle (which is the normal pattern).

Change-Id: Id0e3d1d81b1b166f3c4e843292bdbb597b93acfe