Add extra test for bad url patterns
Report can have parameters from a job substituted into the message.
If a reporter fails for whatever reason it is logged to zuul and
otherwise ignored quietly to the user.

If bad varaibles are attempted to be substituted into a message
log a warning and fallback to the result url for the reporter to
continue.

This was fixed in b7273ef849e7070e21c1d51a9b4190237f385027 so this
change is to just put in an extra test for it.

Change-Id: I521cfbb5873973014c43f9780722d2f80a7c12f0