Cloner: use cache if dest exists
The logic to decide whether or not to use the cache attempted to
detect whether the repo had previously been cloned.  It only did
that by checking whether the destination directory exists.  However,
it's perfectly valid for the dest dir to exist if it is empty.
Adjust the check to look for a .git dir within the dest dir to decide
if the repo has already been cloned.

Change-Id: I17926efcf0f38d6229f0e666e53e6730f455d8ef