Ansible launcher: don't report result on abort
When a worker is asked to abort a job, do not send a result to zuul.
If Zuul asked for the abort, Zuul will ignore the result anyway. If
the operator asked for the abort (i.e., by shutting down the launcher)
then the null result will suggest to Zuul it should relaunch the job.

Change-Id: I8ad35e70f76ede31b78c67ebfb0104554b28e8c2