Ansible launcher: delay node assignment under load
Gearman wakes all available workers at the same time when a job
is available.  The first one to respond gets the job.  To attempt
to more evenly distribute nodes (which are assigned via a gearman
job) across multiple zuul launchers, delay for a period
related exponentially to the number of nodes this launcher currently
has between the wake up and grab job packets.

Change-Id: I307938f97b730b229c1622cd2f929fc5b65ccdad