Enable keep_remote_files in ansible.cfg
When ansible is not using pipelining, files are copied to the remote
worker then run locally on said worker.  Ansible being the good
program that it is, attempts to clean up these files once finished
with them.

As a results, it may be possible for ansible async to delete these
files before actually executing them[1].

Consideration should be taken for long lived workers, as we may want
to update our remote_tmp setting to use the /tmp folder.  We can
discuss that in a follow up patch.

[1] https://github.com/ansible/ansible/issues/10855

Change-Id: Idb1efcdde9d6fe6d74fc2131a4fecddcd1e46904
Signed-off-by: Paul Belanger <pabelanger@redhat.com>