Remove pipeline.getJobs
Previously, building the job tree for a change was the responsibility
of the Pipeline class.  However, with dynamic reconfiguration, that
now lies with QueueItem.  Remove the remaining use of getJobs from
the Pipeline class.  Most of the other invocations of this method
were updated in a previous change.

In doing so, attempt to make the full static reconfiguration handling
compatible with the new dynamic reconfiguration.  This is, as yet,
untested.

Change-Id: I0e40ea36e4834b939a17925f1bb9fd9fa782f07a