Move methods from Pipeline to QueueItem
Because of the previous design where QueueItems were expected to
have a more ephemeral relationship with the Pipelines they were
enqueued into, many actions related to QueueItems were defined
as methods on the Pipeline class.  That is no longer the case
as QueueItems are now firmly situated inside of a pipeline and
it no longer makes sense to ask a Pipeline to perform an action
on an item since all of the data needed to perform that action
is now held by the item itself.

Move all of the methods that operated principally on QueueItems
from Pipeline to QueueItem.

Change-Id: I3250a3fc18537302d856c358a4172735cdf9dca7