Ansible launcher: set remote_tmp
When we use 'delegate_to' to run commands locally, the 'remote'
side of the Ansible connection is the local host.  When running
these tasks it will write to the 'remote_tmp' directory, which
is actually the local ~/.ansible/tmp directory.  We also set
'keep_remote_files' to true in order to avoid a race condition
with 'async' on the actual remote hosts, but in this case, these
two options in combination end up meaning 'keep some files in
the local ~/.ansible/tmp directory indefinitely' which is not
good for our long-running launchers.

Instead, set 'remote_tmp' to a subdirectory of the jobdir so that
when used in the local context, it will be cleaned up at the end
of the run.  In the remote context, it will end up in a similarly
randomly named directory under /tmp on the worker.  Ansible will
create that directory.  This has the side benefit of removing the
Ansible running the job further from potential uses of Ansible
within the job (which may continue to use ~/.ansible by default).

Change-Id: I70475d5844cbd66bf670566f992fdec263d271a5