Ansible launcher: re-register functions after disconnect
Because we are using the private MASS_DO gearman operation to
register functions, the gear.Worker does not know what functions
are registered and therefore the routine which automatically
re-registers functions after a gear server disconnect was not
effective.  Correct this by also storing the function list when
sending MASS_DO.  This will result in the worker actually sending
CAN_DO packets rather than MASS_DO in the case of a reconnect,
but at least it will be correct, if not efficient.

This error would cause existing nodes attached to zuul launchers
to be unable to run jobs after a zuul (geard) restart.

Change-Id: I60804355a8b3a3cfb79a12dd6e6f0e219fe50c31