Tests: ensure fake builds recorded in order
Alter the recording ansible launcher used in tests to record the
build in the launch method which is synchronous, rather than in
the runAnsible method which is run from inside of a thread started
by the launch method.  This way, all builds that show up in the
running_builds attribute appear in the order they arrived from
gearman (which is the order in which they were launched).

Change-Id: I11d8c686f738b51797f9ac9cee0ac201800de383