Re-enable test_failed_change_at_head
And "improve" it to use the new build/history assertions.

Add an "ordered" option to assertHistory so that we can assert
everything about the history except that the builds arrived in
the specified order.  In this case, aborted builds don't always
finish in order.

Also add the ordered option to test_failed_changes since it
is subject to the same issue.

Change-Id: I7b3bec798b462568d4c44db8943daaeb27728735