bump timeout on ssh commands to 30s
We are seeing intermitent failures in zuul trying to talk to the node
which look like they are the 10s ssh negotiation failing. Extremely
busy test nodes that are using their entire network bw to pull
packages, may take longer than this.

Try to reduce this by bumping the timeout.

Change-Id: Ic4ec2ea3c8b77cb308fb1a85514d831acf6c4b67