Delete test_single_nonexistent_post_job test
This is basically testing the case where Zuul wants to tell Jenkins to
run a job, but Jenkins doesn't know about it.  This doesn't really
have an analog in a Zuul v3 world, where we no longer have that
division.  Adding a job to a pipeline with no job definition is really
just a configuration error that we should catch in validation.

Change-Id: I8f9cc063e0bd6472fbc78740279f51280f34ea64
Signed-off-by: Paul Belanger <pabelanger@redhat.com>