Don't merge post-merge items
In all cases, the launcher-merger updates the repos involved in a
job before running it.  If there are pre-merge changes, it then
merges those changes into the repos.  If the job does not involve
pre-merge changes, then nothing further needs to happen.  Avoid
attempting to merge changes which are already merged in this case.

Change-Id: Ie0c0d258b4edad4afc3b569f8ea222523bc769c1