Read all Gerrit events from poll interruption
Gerrit may emits multiple events in a single packet, in particular the
change-created and ref-updated events. This results in a single poll
interruption and the GerritWatcher will only process the first event
until there is more activity on the event streams.

To correct this, the _read method needs to read all the lines already
available in the paramiko channel file before polling the file.

Story: 2000816
Task: 3401
Change-Id: I93b55a6491353aee8c954d349cef658c09d1ff2d