Re-enable TestScheduler.test_rerun_on_error
This test is making sure that when something terrible goes wrong in the
launcher and it just returns 'None', the scheduler retries.

During refactoring of the launcher, some of the logic that handled this
special case of run_error was removed. Also the launcher wasn't properly
handling a return of None and would have failed with a TypeError instead
of sending a failure to the zuul client.

Change-Id: I6b063ba913bf72087d2cc027f08e02304310c2be
Story: 2000773
Task: 3403