Re-enable test_rerun_on_abort
The order of operations for this one has changed a bit due to the
different interactions with launchers. We run one less operation now
because the first release actually does do the full launch-dance and
incur a single retry before we enter the loop of retries. Because of
that, our history is one less, and the number of times we force retry is
one less.

Change-Id: I6d7d8d0f7e49a11e926667fbe772535ebdd35e89
Story: 2000773
Task: 3413