Re-enable test_client_get_running_jobs
In coercing this test to run, it's clear there are a number of "TODO"
items in zuul.launcher.server.LaunchServer that seem simple enough but
that I don't really understand. I've filled in enough to make the test
pass, but I am not at all confident that this is actually the way we
want this to work long-term.

get_running_jobs also had to be extended to add tenants. I have to
wonder if we should change the payload to return the tenants somehow.

Change-Id: If91cb662ceef8e1d7660974df07b821720f210d4
Story: 2000773
Task: 3414