Fix test_rerun_on_abort
Change I6d7d8d0f7e49a11e926667fbe772535ebdd35e89 erroneously altered
test_rerun_on_abort to match the observed behavior.

Change I6e64ef03cbb10ce858b22d6a4590f58ace0a5332 restored the values
in the test but then erroneously changed the accounting system to
match the observed behavior.

The actual problem is that this test also exercises the 'attempts' job
attribute, and does so by specifying it in a custom test configuration.
The test was failing because that attribute, which instructs zuul to
attempt to run a job for a non-default number of retries, was not being
read.

It was not being read because the test was still using the old configuration
loading scheme.  It updated the "layout_file" which isn't a thing anymore
and asked the scheduler to reload.  The scheduler *did* reload, but it
simply reloaded the same configuration.  The solution to this is to either
create a new configuration, or, in this case, since the additional
configuration needed is compatible with the configuration used by the tests
siblings, simply add it to the active config file for the test.

Once the test is loading the correct configuration, one can observe that
the 'attempts' attribute was not added to the validator.  That is corrected
in this change as well.

With all of this complete, the test passes in its original form and no
modifications to the job retry accounting system.

Change-Id: Icf6d697cbae0166bc516faf5b7e60cac05885ab0