Re-enable test_dependent_behind_dequeue
The number of pipelines and thus, jobs, that run is less in this new v3
common configuration than it was in master. Notably, in v3, we
only see 'check', 'gate', 'experimental', and 'post'. But in master we
also have dup1, dup2, conflict, and unused.

Meanwhile, it appears that the jobs that should abort, do, and the ones
that should not, do not.

Change-Id: Iefb257ea6477d504c78d40dd411ac9eaf3125ae5
Story: 2000773
Task: 3436